Move docs to standalone module
