+++ lib/docs/admin/architecture.md
