+++ lib/docs/admin/components.md
