+++ lib/docs/assets/LisaArchitecture.png
