+++ lib/docs/config/branding.md
