+++ lib/docs/config/hiding-chat-components.md
