+++ lib/docs/config/vector-stores.md
