+++ lib/docs/public/favicon.ico
