+++ lib/docs/public/logo.png
