+++ lib/docs/user/rag.md
