+++ lib/user-interface/react/package.json
@@ -5,13 +5,10 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "node ./scripts/set-revision-info.mjs && tsc && vite build && npm run docs:build",
+    "build": "node ./scripts/set-revision-info.mjs && tsc && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
-    "format": "prettier --ignore-path .gitignore --write \"**/*.+(tsx|js|ts|json)\"",
-    "docs:dev": "vitepress dev docs",
-    "docs:build": "vitepress build docs",
-    "docs:preview": "vitepress preview docs"
+    "format": "prettier --ignore-path .gitignore --write \"**/*.+(tsx|js|ts|json)\""
   },
   "dependencies": {
     "@cloudscape-design/collection-hooks": "^1.0.23",
@@ -60,7 +57,6 @@
     "prettier": "^3.0.0",
     "redux-mock-store": "^1.5.4",
     "uuid": "^9.0.0",
-    "vite": "^4.5.3",
-    "vitepress": "^1.4.1"
+    "vite": "^4.5.3"
   }
 }
